requirements.txt
Django==2.1.5
gunicorn==19.9.0
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
beautifulsoup4==4.7.1
